Reformatted the Play view + implemented the timer fully
So I reformatted the play view to incorprorate 3 buttons for testing purposes, and the usecases for timer and score seems to work well, and can be call by function in controller. I also modified several enitites to make this work. Lastly, the timer for counting down is ran by a threaded function in the timerInteractor.